chore: move fetch models to client
