+++ apps/web/app/preview/whisper/components/data-table.tsx
@@ -36,6 +36,7 @@ import { Autostart } from "./Autostart";
 import { LanguageEnum, ModelType, TableType } from "../data/types";
 import { Task } from "../data/schema";
 import { th } from "@faker-js/faker";
+import { useModels } from "./hooks/useModels";
 
 const socket = io(baseURL);
 
@@ -46,7 +47,6 @@ socket.on("connection", (message) => {
 interface DataTableProps<TData, TValue> {
   columns: ColumnDef<TData, TValue>[];
   data?: TData[];
-  models: ModelType[];
   type: TableType;
 }
 
@@ -108,10 +108,10 @@ function useList(type: TableType) {
   };
 }
 
+
 export function DataTable<TData extends Task, TValue>({
   columns,
   data: initData = [],
-  models,
   type,
 }: DataTableProps<TData, TValue>) {
   const [rowSelection, setRowSelection] = React.useState({});
@@ -122,6 +122,7 @@ export function DataTable<TData extends Task, TValue>({
   );
   const [sorting, setSorting] = React.useState<SortingState>([]);
   const { list } = useList(type);
+  const { data: models = [], isLoading: modelsLoading } = useModels();
   const [model, setModel] = React.useState<ModelType | undefined>(models?.[0]);
 
   const [data, setData] = React.useState(initData);
